Solve problems by comments.
